requirements.txt
docker==4.3.1
biopython==1.78
varname==0.4.0